Cleaned up GL a minor amount more. Sounds can now be recorded while capturing to tga sequences (as raw sound). Fixed frame limits a little. Passed the unicode key values around to a few more key input functions.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@3332 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/gl/gl_vidnt.c
@@ -435,7 +435,7 @@ qboolean VID_SetWindowedMode (rendererstate_t *info)
 				SetWindowLong(dibwindow, GWL_EXSTYLE, GetWindowLong(dibwindow, GWL_EXSTYLE) | WS_EX_LAYERED);
 
 				// Make this window 70% alpha
-				pSetLayeredWindowAttributes(dibwindow, 0, av, LWA_ALPHA);
+				pSetLayeredWindowAttributes(dibwindow, 0, (BYTE)av, LWA_ALPHA);
 			}
 		}
 #endif
@@ -1031,7 +1031,7 @@ void VID_WndAlpha_Override_Callback(struct cvar_s *var, char *oldvalue)
 				SetWindowLong(mainwindow, GWL_EXSTYLE, GetWindowLong(mainwindow, GWL_EXSTYLE) | WS_EX_LAYERED);
 
 				// Make this window 70% alpha
-				pSetLayeredWindowAttributes(mainwindow, 0, av, LWA_ALPHA);
+				pSetLayeredWindowAttributes(mainwindow, 0, (BYTE)av, LWA_ALPHA);
 			}
 			else
 				SetWindowLong(mainwindow, GWL_EXSTYLE, GetWindowLong(mainwindow, GWL_EXSTYLE) & ~WS_EX_LAYERED);
